Move dma to channel pass from Conversion to Transform
