+++ mlir/include/air/Conversion/PassDetail.h
@@ -28,7 +28,6 @@ using namespace mlir;
 #define GEN_PASS_DEF_AIRTOAIE
 #define GEN_PASS_DEF_AIRTOASYNC
 #define GEN_PASS_DEF_COPYTODMA
-#define GEN_PASS_DEF_DMATOCHANNEL
 #define GEN_PASS_DEF_INSERTEMPTYLAUNCHOVERHERD
 #define GEN_PASS_DEF_PARALLELTOHERD
 #define GEN_PASS_DEF_PARALLELTOLAUNCH
